+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 Breaking changes
 
+- cmd/bnsd: `nft/username` allows now for any number of aliases/names for a
+  single address.
 - messages produced by `cmd/bnscli` have a new binary format incompatible with
   the previous version.
 - `x/gov` added indexes to proposals and electorate to enable better client-side UX